refactor: make tests pass again
