+++ lib/adapters/http.js
@@ -182,6 +182,7 @@ export default function httpAdapter (config) {
         settle(resolve, reject, response)
       } else {
         const responseBuffer = []
+        response = responsePartial
         stream.on('data', function handleStreamData (chunk) {
           responseBuffer.push(chunk)
