+++ lib/axios.js
@@ -1,10 +1,6 @@
-'use strict'
-
-const utils = require('./utils')
-const bind = require('lodash/bind')
-const Axios = require('./core/Axios')
-const mergeConfig = require('./core/mergeConfig')
-const defaults = require('./defaults')
+import Axios from './core/Axios'
+import mergeConfig from './core/mergeConfig'
+import defaults from './defaults'
 
 /**
  * Create an instance of Axios
@@ -13,41 +9,25 @@ const defaults = require('./defaults')
  * @return {Axios} A new instance of Axios
  */
 function createInstance (defaultConfig) {
-  const context = new Axios(defaultConfig)
-  const instance = bind(Axios.prototype.request, context)
-
-  // Copy axios.prototype to instance
-  utils.extend(instance, Axios.prototype, context)
-
-  // Copy context to instance
-  utils.extend(instance, context)
-
-  return instance
+  return new Axios(defaultConfig)
 }
 
 // Create the default instance to be exported
 const axios = createInstance(defaults)
-
-// Expose Axios class to allow class inheritance
-axios.Axios = Axios
 
 // Factory for creating new instances
 axios.create = function create (instanceConfig) {
   return createInstance(mergeConfig(axios.defaults, instanceConfig))
 }
 
+export default axios
+
 // Expose Cancel & CancelToken
-axios.Cancel = require('./cancel/Cancel')
-axios.CancelToken = require('./cancel/CancelToken')
-axios.isCancel = require('./cancel/isCancel')
+export { default as Cancel } from './cancel/Cancel'
+export { default as CancelToken } from './cancel/CancelToken'
+export { default as isCancel } from './cancel/isCancel'
 
-// Expose all/spread
-axios.all = function all (promises) {
-  return Promise.all(promises)
+export {
+  // Expose Axios class to allow class inheritance
+  Axios
 }
-axios.spread = require('./helpers/spread')
-
-module.exports = axios
-
-// Allow use of default import syntax in TypeScript
-module.exports.default = axios
