+++ lib/browser.js
@@ -2,8 +2,8 @@
 const axios = require('./axios')
 const defaultInstance = axios.default
 
-delete axios.default
-
 Object.assign(defaultInstance, axios)
+
+delete defaultInstance.default
 
 module.exports = defaultInstance
