+++ lib/core/dispatchRequest.js
@@ -1,7 +1,7 @@
 import { merge, forEach } from 'lodash'
 import transformData from './transformData'
+import defaults from '../defaults'
 const isCancel = require('../cancel/isCancel')
-const defaults = require('../defaults')
 const isAbsoluteURL = require('./../helpers/isAbsoluteURL')
 const combineURLs = require('./../helpers/combineURLs')
