+++ test/specs/transform.spec.js
@@ -1,6 +1,6 @@
 /* global getAjaxRequest */
 
-import axios from '../../index'
+import axios from '../../lib/axios'
 
 describe('transform', function () {
   beforeEach(function () {
